requirements.txt
django>=5.1,<6.0
django-environ>=0.11
whitenoise>=6.6
django-jazzmin>=3.0
Pillow>=10.0
pdf2image>=1.17
